(PC-21318)[PRO] fix: move address validation test
