+++ pro/src/components/VenueForm/validationSchema.ts
@@ -11,11 +11,13 @@ const validationSchemaConcat = (newOnboardingActive: boolean) => {
     ...activitySchema(newOnboardingActive),
     ...accessibilityValidationSchema,
     ...contactValidationSchema,
+    isVenueVirtual: yup.boolean(),
     addressAutocomplete: yup.string().when('isVenueVirtual', {
       is: false,
-      then: yup
-        .string()
-        .required('Veuillez sélectionner une adresse parmi les suggestions'),
+      then: schema =>
+        schema.required(
+          'Veuillez sélectionner une adresse parmi les suggestions'
+        ),
     }),
   }
 }
