+++ pro/src/screens/SignupJourneyForm/Authentication/validationSchema.tsx
@@ -1,10 +1,10 @@
 import * as yup from 'yup'
 
+import { validationSchema as addressValidationSchema } from 'components/Address'
+
 export const validationSchema = yup.object().shape({
   siret: yup.string().required(),
   name: yup.string().required(),
   publicName: yup.string().nullable(),
-  addressAutocomplete: yup
-    .string()
-    .required('Veuillez sélectionner une adresse parmi les suggestions'),
+  ...addressValidationSchema,
 })
